backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
requests
joblib
lightgbm
scikit-learn
pandas
numpy
tqdm
python-multipart
